first pass at sharing code between tensorlib and cwrap
