+++ tools/shared/__init__.py
@@ -1,0 +1,1 @@
+from .module_loader import import_module
